// src/types/UserStage.ts
export enum UserStage {
  WAITING_APPLICATION_TRAINING = 'WAITING_APPLICATION_TRAINING',
  REFUSAL_APPLICATION = 'REFUSAL_APPLICATION',
  PASSES_ENTRANCE_TEST = 'PASSES_ENTRANCE_TEST',
  FAILED_ENTRANCE_TEST = 'FAILED_ENTRANCE_TEST',
  STUDYING = 'STUDYING',
  FAILED_STUDYING = 'FAILED_STUDYING',
  EXPECTS_PRODUCTION_PRACTICE = 'EXPECTS_PRODUCTION_PRACTICE',
  PRODUCTION_PRACTICE = 'PRODUCTION_PRACTICE',
  EXAM = 'EXAM',
  GRADUATED = 'GRADUATED',
  FAILED_EXAM = 'FAILED_EXAM'
}